convert gamma tests
